WIP: T04 -- Add tests for untested UNION methods
