+++ tests/integration/telemetry/test_telemetry_query_service_misc.py
@@ -1,4 +1,8 @@
-"""Integration tests for TelemetryQueryService — source tier, job summary, health, and activity feed."""
+"""Integration tests for TelemetryQueryService — source tier, job summary, and health.
+
+UNION-based methods (get_app_recent_activity, get_per_app_activity_buckets,
+get_per_app_last_errors) live in test_union_queries.py.
+"""
 
 import asyncio
 import time
@@ -13,7 +17,6 @@ from hassette.core.database_service import DatabaseService
 from hassette.core.telemetry.helpers import source_tier_clause
 from hassette.core.telemetry.query_service import TelemetryQueryService
 from hassette.exceptions import TelemetryUnavailableError
-from hassette.schemas.execution_models import ActivityFeedEntry
 from hassette.test_utils.mock_hassette import make_mock_hassette
 
 from .helpers import (
@@ -234,285 +237,3 @@ class TestReadTimeout:
     ) -> None:
         """A fast query completes within even a short timeout."""
         await short_timeout_query_service.check_health()
-
-
-class TestGetAppRecentActivity:
-    async def test_merged_sorted_by_timestamp_desc(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """Handler invocations and job executions are merged and sorted by timestamp DESC."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        listener_id = await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-        job_id = await insert_job(db_svc, app_key="test_app", job_name="my_job", handler_method="run_job")
-
-        # Interleave timestamps so merge order is testable
-        await insert_invocation(db_svc, listener_id, session_id, status="success", execution_start_ts=base_ts + 30.0)
-        await insert_invocation(
-            db_svc, listener_id, session_id, status="error", execution_start_ts=base_ts + 10.0, error_type="ValueError"
-        )
-        await insert_execution(db_svc, job_id, session_id, status="success", execution_start_ts=base_ts + 20.0)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 3
-        assert all(isinstance(r, ActivityFeedEntry) for r in results)
-        # Sorted DESC by timestamp
-        assert results[0].timestamp == pytest.approx(base_ts + 30.0)
-        assert results[1].timestamp == pytest.approx(base_ts + 20.0)
-        assert results[2].timestamp == pytest.approx(base_ts + 10.0)
-
-    async def test_kind_field_correct(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """Handler invocations have kind='handler', job executions have kind='job'."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        listener_id = await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-        job_id = await insert_job(db_svc, app_key="test_app", job_name="my_job", handler_method="run_job")
-
-        await insert_invocation(db_svc, listener_id, session_id, status="success", execution_start_ts=base_ts + 20.0)
-        await insert_execution(db_svc, job_id, session_id, status="success", execution_start_ts=base_ts + 10.0)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 2
-        assert results[0].kind == "handler"
-        assert results[0].handler_id == listener_id
-        assert results[1].kind == "job"
-        assert results[1].handler_id == job_id
-
-    async def test_limit_is_respected(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """Limit parameter caps the number of returned entries."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        listener_id = await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-
-        for i in range(10):
-            await insert_invocation(
-                db_svc, listener_id, session_id, status="success", execution_start_ts=base_ts + float(i)
-            )
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=3,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 3
-        # Should be the 3 most recent
-        assert results[0].timestamp == pytest.approx(base_ts + 9.0)
-        assert results[1].timestamp == pytest.approx(base_ts + 8.0)
-        assert results[2].timestamp == pytest.approx(base_ts + 7.0)
-
-    async def test_since_filters_old_entries(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """Since parameter excludes entries older than the threshold."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        since_ts = base_ts + 15.0
-
-        listener_id = await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-        job_id = await insert_job(db_svc, app_key="test_app", job_name="my_job", handler_method="run_job")
-
-        # After since_ts — should be included
-        await insert_invocation(db_svc, listener_id, session_id, status="success", execution_start_ts=base_ts + 20.0)
-        await insert_execution(db_svc, job_id, session_id, status="success", execution_start_ts=base_ts + 30.0)
-
-        # Before since_ts — should be excluded
-        await insert_invocation(db_svc, listener_id, session_id, status="error", execution_start_ts=base_ts + 5.0)
-        await insert_execution(db_svc, job_id, session_id, status="error", execution_start_ts=base_ts + 10.0)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=since_ts,
-            source_tier="app",
-        )
-
-        assert len(results) == 2
-        assert all(r.timestamp >= since_ts for r in results)
-
-    async def test_source_tier_filtering(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """source_tier='framework' returns only framework-tier entries, not app-tier."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        app_listener = await insert_listener(db_svc, app_key="test_app", handler_method="on_app", source_tier="app")
-        fw_listener = await insert_listener(db_svc, app_key="test_app", handler_method="on_fw", source_tier="framework")
-
-        await insert_invocation(
-            db_svc, app_listener, session_id, status="success", execution_start_ts=base_ts + 10.0, source_tier="app"
-        )
-        await insert_invocation(
-            db_svc,
-            fw_listener,
-            session_id,
-            status="success",
-            execution_start_ts=base_ts + 20.0,
-            source_tier="framework",
-        )
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="framework",
-        )
-
-        assert len(results) == 1
-        assert results[0].handler_name == "on_fw"
-
-    async def test_instance_index_scoping(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """instance_index filters to entries for that instance only."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        listener_0 = await insert_listener(db_svc, app_key="test_app", instance_index=0, handler_method="on_event")
-        listener_1 = await insert_listener(db_svc, app_key="test_app", instance_index=1, handler_method="on_event")
-
-        await insert_invocation(db_svc, listener_0, session_id, status="success", execution_start_ts=base_ts + 10.0)
-        await insert_invocation(db_svc, listener_1, session_id, status="success", execution_start_ts=base_ts + 20.0)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=0,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 1
-        assert results[0].timestamp == pytest.approx(base_ts + 10.0)
-
-    async def test_empty_app_returns_empty_list(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """App with no invocations or executions returns an empty list."""
-        db_svc, _session_id = db
-        await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert results == []
-
-    async def test_isolates_to_app_key(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """Results are scoped to the requested app_key only."""
-        db_svc, session_id = db
-
-        base_ts = BASE_TS
-        listener_a = await insert_listener(db_svc, app_key="app_a", handler_method="on_a")
-        listener_b = await insert_listener(db_svc, app_key="app_b", handler_method="on_b")
-
-        await insert_invocation(db_svc, listener_a, session_id, status="success", execution_start_ts=base_ts + 10.0)
-        await insert_invocation(db_svc, listener_b, session_id, status="success", execution_start_ts=base_ts + 20.0)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="app_a",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 1
-        assert results[0].app_key == "app_a"
-
-    async def test_row_id_uniqueness_and_prefixes(
-        self,
-        query_service: TelemetryQueryService,
-        db: tuple[DatabaseService, int],
-    ) -> None:
-        """row_id values are unique across all rows and use the correct kind prefix."""
-        db_svc, session_id = db
-
-        # Same timestamp for both invocations to stress-test uniqueness
-        shared_ts = 1_000_000.0
-
-        listener_id = await insert_listener(db_svc, app_key="test_app", handler_method="on_event")
-        job_id = await insert_job(db_svc, app_key="test_app", job_name="my_job", handler_method="run_job")
-
-        # Two handler invocations with the same timestamp
-        await insert_invocation(db_svc, listener_id, session_id, status="success", execution_start_ts=shared_ts)
-        await insert_invocation(
-            db_svc, listener_id, session_id, status="error", execution_start_ts=shared_ts, error_type="ValueError"
-        )
-        # One job execution with the same timestamp
-        await insert_execution(db_svc, job_id, session_id, status="success", execution_start_ts=shared_ts)
-
-        results = await query_service.get_app_recent_activity(
-            app_key="test_app",
-            instance_index=None,
-            limit=50,
-            since=None,
-            source_tier="app",
-        )
-
-        assert len(results) == 3
-
-        # All row_id values must be present and unique
-        row_ids = [r.row_id for r in results]
-        assert len(set(row_ids)) == 3, f"Expected 3 unique row_ids, got: {row_ids}"
-
-        # Handler rows prefixed with 'h-', job rows with 'j-'
-        handler_rows = [r for r in results if r.kind == "handler"]
-        job_rows = [r for r in results if r.kind == "job"]
-
-        assert len(handler_rows) == 2
-        assert len(job_rows) == 1
-
-        for r in handler_rows:
-            assert r.row_id.startswith("h-"), f"Handler row_id should start with 'h-', got: {r.row_id!r}"
-        for r in job_rows:
-            assert r.row_id.startswith("j-"), f"Job row_id should start with 'j-', got: {r.row_id!r}"
